backend/requirements.txt
fastapi==0.115.2
uvicorn==0.32.0
gunicorn==23.0.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
pgvector==0.3.6
pymupdf==1.24.10
python-docx==1.1.2
sentence-transformers==3.1.1
numpy==2.1.2
requests==2.32.3
pydantic==2.9.2
pydantic-settings==2.6.0
python-multipart==0.0.9
python-dotenv==1.0.1
pytest==8.3.3
httpx==0.27.2
